refactor: move all webview version stuff to WebViewUtils
