+++ AnkiDroid/src/main/java/com/ichi2/anki/InitialActivity.kt
@@ -18,17 +18,13 @@ package com.ichi2.anki
 
 import android.content.Context
 import android.content.SharedPreferences
-import android.content.pm.PackageInfo
-import android.content.pm.PackageManager
 import android.database.sqlite.SQLiteFullException
 import android.os.Build
 import android.os.Environment
 import android.os.Parcelable
 import androidx.annotation.CheckResult
 import androidx.annotation.RequiresApi
-import androidx.appcompat.app.AlertDialog
 import androidx.core.content.edit
-import androidx.core.content.pm.PackageInfoCompat
 import com.ichi2.anki.exception.StorageAccessException
 import com.ichi2.anki.servicelayer.PreferenceUpgradeService
 import com.ichi2.anki.servicelayer.PreferenceUpgradeService.setPreferencesUpToDate
@@ -37,11 +33,8 @@ import com.ichi2.anki.ui.windows.permissions.Full30and31PermissionsFragment
 import com.ichi2.anki.ui.windows.permissions.PermissionsFragment
 import com.ichi2.anki.ui.windows.permissions.PermissionsUntil29Fragment
 import com.ichi2.anki.ui.windows.permissions.TiramisuPermissionsFragment
-import com.ichi2.utils.OLDEST_WORKING_WEBVIEW_VERSION_CODE
 import com.ichi2.utils.Permissions
 import com.ichi2.utils.VersionUtils.pkgVersionName
-import com.ichi2.utils.getAndroidSystemWebViewPackageInfo
-import com.ichi2.utils.show
 import kotlinx.parcelize.Parcelize
 import net.ankiweb.rsdroid.BackendException
 import timber.log.Timber
@@ -145,57 +138,6 @@ object InitialActivity {
     enum class StartupFailure {
         SD_CARD_NOT_MOUNTED, DIRECTORY_NOT_ACCESSIBLE, FUTURE_ANKIDROID_VERSION,
         DB_ERROR, DATABASE_LOCKED, WEBVIEW_FAILED, DISK_FULL
-    }
-
-    /**
-     * Shows a dialog if the current WebView version is older than the last supported version.
-     */
-    fun checkWebviewVersion(packageManager: PackageManager, activity: AnkiActivity) {
-        val webviewPackageInfo = getAndroidSystemWebViewPackageInfo(packageManager) ?: return
-        val webviewVersion = webviewPackageInfo.versionName ?: run {
-            Timber.w("Failed to obtain WebView version")
-            return
-        }
-        val versionCode = PackageInfoCompat.getLongVersionCode(webviewPackageInfo)
-        // TODO modify the alert dialog text to handle the usage of developer builds for system WebView
-        val userVisibleCode = runCatching {
-            webviewVersion.split(".")[0].toInt()
-        }.getOrNull() ?: 0
-        if (versionCode >= OLDEST_WORKING_WEBVIEW_VERSION_CODE) {
-            Timber.d(
-                "WebView is up to date. %s: %s(%s)",
-                webviewPackageInfo.packageName,
-                webviewVersion,
-                versionCode.toString()
-            )
-            return
-        }
-
-        val legacyWebViewPackageInfo = getLegacyWebViewPackageInfo(packageManager)
-        if (legacyWebViewPackageInfo != null) {
-            Timber.w("WebView is outdated. %s: %s", legacyWebViewPackageInfo.packageName, legacyWebViewPackageInfo.versionName)
-            showOutdatedWebViewDialog(activity, userVisibleCode, activity.getString(R.string.link_legacy_webview_update))
-        } else {
-            Timber.w("WebView is outdated. %s: %s", webviewPackageInfo.packageName, webviewPackageInfo.versionName)
-            showOutdatedWebViewDialog(activity, userVisibleCode, activity.getString(R.string.link_webview_update))
-        }
-    }
-
-    private fun showOutdatedWebViewDialog(activity: AnkiActivity, installedVersion: Int, learnMoreUrl: String) {
-        AlertDialog.Builder(activity).show {
-            setMessage(activity.getString(R.string.webview_update_message, installedVersion, OLDEST_WORKING_WEBVIEW_VERSION))
-            setPositiveButton(R.string.scoped_storage_learn_more) { _, _ ->
-                activity.openUrl(learnMoreUrl)
-            }
-        }
-    }
-
-    private fun getLegacyWebViewPackageInfo(packageManager: PackageManager): PackageInfo? {
-        return try {
-            packageManager.getPackageInfo("com.android.webview", 0)
-        } catch (e: PackageManager.NameNotFoundException) {
-            null
-        }
     }
 }
