+++ AnkiDroid/src/main/java/com/ichi2/utils/WebViewUtils.kt
@@ -16,15 +16,86 @@
 
 package com.ichi2.utils
 
+import android.content.Context
 import android.content.pm.PackageInfo
 import android.content.pm.PackageManager
+import android.webkit.WebView
+import androidx.annotation.MainThread
+import androidx.appcompat.app.AlertDialog
+import androidx.core.content.pm.PackageInfoCompat
+import com.ichi2.anki.AnkiActivity
+import com.ichi2.anki.CrashReportService
+import com.ichi2.anki.R
+import timber.log.Timber
 
 internal const val OLDEST_WORKING_WEBVIEW_VERSION_CODE = 386507305L
+internal const val OLDEST_WORKING_WEBVIEW_VERSION = 77
+
+/**
+ * Shows a dialog if the current WebView version is older than the last supported version.
+ */
+fun checkWebviewVersion(packageManager: PackageManager, activity: AnkiActivity) {
+    val webviewPackageInfo = getAndroidSystemWebViewPackageInfo(packageManager) ?: return
+    val webviewVersion = webviewPackageInfo.versionName ?: run {
+        Timber.w("Failed to obtain WebView version")
+        return
+    }
+    val versionCode = PackageInfoCompat.getLongVersionCode(webviewPackageInfo)
+    // TODO modify the alert dialog text to handle the usage of developer builds for system WebView
+    val userVisibleCode = runCatching {
+        webviewVersion.split(".")[0].toInt()
+    }.getOrNull() ?: 0
+    if (versionCode >= OLDEST_WORKING_WEBVIEW_VERSION_CODE) {
+        Timber.d(
+            "WebView is up to date. %s: %s(%s)",
+            webviewPackageInfo.packageName,
+            webviewVersion,
+            versionCode.toString()
+        )
+        return
+    }
+
+    val legacyWebViewPackageInfo = getLegacyWebViewPackageInfo(packageManager)
+    if (legacyWebViewPackageInfo != null) {
+        Timber.w("WebView is outdated. %s: %s", legacyWebViewPackageInfo.packageName, legacyWebViewPackageInfo.versionName)
+        showOutdatedWebViewDialog(activity, userVisibleCode, activity.getString(R.string.link_legacy_webview_update))
+    } else {
+        Timber.w("WebView is outdated. %s: %s", webviewPackageInfo.packageName, webviewPackageInfo.versionName)
+        showOutdatedWebViewDialog(activity, userVisibleCode, activity.getString(R.string.link_webview_update))
+    }
+}
+
+@MainThread
+fun getWebviewUserAgent(context: Context): String? {
+    try {
+        return WebView(context).settings.userAgentString
+    } catch (e: Throwable) {
+        CrashReportService.sendExceptionReport(e, "WebViewUtils", "some issue occurred while extracting webview user agent")
+    }
+    return null
+}
+
+private fun showOutdatedWebViewDialog(activity: AnkiActivity, installedVersion: Int, learnMoreUrl: String) {
+    AlertDialog.Builder(activity).show {
+        setMessage(activity.getString(R.string.webview_update_message, installedVersion, OLDEST_WORKING_WEBVIEW_VERSION))
+        setPositiveButton(R.string.scoped_storage_learn_more) { _, _ ->
+            activity.openUrl(learnMoreUrl)
+        }
+    }
+}
+
+private fun getLegacyWebViewPackageInfo(packageManager: PackageManager): PackageInfo? {
+    return try {
+        packageManager.getPackageInfo("com.android.webview", 0)
+    } catch (_: PackageManager.NameNotFoundException) {
+        null
+    }
+}
 
 /**
  * Returns a [PackageInfo] from the current system WebView, or `null` if unavailable
  */
-fun getAndroidSystemWebViewPackageInfo(packageManager: PackageManager): PackageInfo? {
+private fun getAndroidSystemWebViewPackageInfo(packageManager: PackageManager): PackageInfo? {
     fun getPackage(packageName: String): PackageInfo? {
         return try {
             packageManager.getPackageInfo(packageName, 0)
